replaced crewai agents with autogen
