+++ pyproject.toml
@@ -7,7 +7,6 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.13"
-crewai-tools = "0.1.1"
 pyright = "^1.1.350"
 click = "^8.1.7"
 safe-eth-py = "^5.8.0"
@@ -15,7 +14,6 @@ eth-account = "^0.11.0"
 web3 = "^6.15.1"
 requests = "^2.31.0"
 python-dotenv = "^1.0.0"
-crewai = "0.22.5"
 coingecko = "^0.13"
 pyautogen = "^0.2.21"
